'--overlap' as float.

+++ CookHLA.py
@@ -581,7 +581,7 @@ if __name__ == "__main__":
 
     # Beagle5.1.
     parser.add_argument("--overlap", "-ol",
-                        help="\n3 Overlap values(cM) for Beagle 5.1 implementation.\n\n", nargs=3, default=(4,8,12), type=int)
+                        help="\n3 Overlap values(cM) for Beagle 5.1 implementation.\n\n", nargs=3, default=(4,8,12), type=float)
     parser.add_argument("--window", "-w",
                         help="\nwindow values(cM) for Beagle 5.1 implementation.\n\n", default=40, type=int)
